storage: break dep of mz-storage-types on mz-persist-txn
Pure refactor, no behavior change.

Accomplished by moving `TxnsEntry` and `TxnsCodec` into a new module.
