+++ ci/test/lint-deps.toml
@@ -92,6 +92,12 @@ dependencies = [
     "mz-persist-client",
 ]
 
+[[deny]]
+name = "mz-storage-types"
+dependencies = [
+    "mz-persist-txn",
+]
+
 # Keep various crates in only one of environmentd or clusterd.
 [[deny]]
 name = "mz-environmentd"
