+++ src/transform/Cargo.toml
@@ -16,7 +16,6 @@ itertools = "0.10.5"
 mz-compute-types = { path = "../compute-types" }
 mz-expr = { path = "../expr" }
 mz-ore = { path = "../ore" }
-mz-persist-client = { path = "../persist-client" }
 mz-repr = { path = "../repr", features = ["tracing_"] }
 num-traits = "0.2"
 ordered-float = { version = "4.2.0", features = ["serde"] }
